Restructure Examples and add e2e tests
