+++ packages/swap/example/Clickable/index.js
@@ -1,0 +1,1 @@
+export * from './Clickable';
