+++ packages/swap/example/Deep/index.js
@@ -1,0 +1,1 @@
+export * from './Deep';
