+++ packages/swap/example/Hoverable/index.js
@@ -1,0 +1,1 @@
+export * from './Hoverable';
