+++ packages/swap/example/Off/Off.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export function Off({children, ...rest}) {
+  return (
+    <div data-e2e="off" className="off" {...rest}>
+      {children}
+    </div>
+  );
+}
+Off.propTypes = {
+  children: PropTypes.node
+};
+Off.defaultProps = {
+  children: 'OFF'
+};
