+++ packages/swap/example/ToggleFromOutside/index.js
@@ -1,0 +1,1 @@
+export * from './ToggleFromOutside';
