--- packages/swap/example/app.css
@@ -1,22 +1,0 @@
-.app {
-  padding: 2em;
-}
-
-.on {
-  width: 100px;
-  height: 100px;
-  border-radius: 50px;
-  background-color: #c00;
-  color: #fff;
-  text-align: center;
-  line-height: 100px;
-}
-
-.off {
-  width: 100px;
-  height: 100px;
-  background-color: #06c;
-  color: #fff;
-  text-align: center;
-  line-height: 100px;
-}
